address PR comments & cleanup

+++ consensus/testing.go
@@ -49,7 +49,7 @@ func RequireTipSetAdd(require *require.Assertions, blk *types.Block, ts TipSet) 
 }
 
 // TestPowerTableView is an implementation of the powertable view used for testing mining
-// wherein each miner has 1/n power.
+// wherein each miner has totalPower/minerPower power.
 type TestPowerTableView struct{ minerPower, totalPower uint64 }
 
 // NewTestPowerTableView creates a test power view with the given total power
@@ -57,14 +57,17 @@ func NewTestPowerTableView(minerPower uint64, totalPower uint64) *TestPowerTable
 	return &TestPowerTableView{minerPower, totalPower}
 }
 
+// Total always returns value that was supplied to NewTestPowerTableView.
 func (tv *TestPowerTableView) Total(ctx context.Context, st state.Tree, bstore blockstore.Blockstore) (uint64, error) {
 	return tv.totalPower, nil
 }
 
+// Miner always returns value that was supplied to NewTestPowerTableView.
 func (tv *TestPowerTableView) Miner(ctx context.Context, st state.Tree, bstore blockstore.Blockstore, mAddr address.Address) (uint64, error) {
-	return uint64(tv.minerPower), nil
+	return tv.minerPower, nil
 }
 
+// HasPower always returns true.
 func (tv *TestPowerTableView) HasPower(ctx context.Context, st state.Tree, bstore blockstore.Blockstore, mAddr address.Address) bool {
 	return true
 }
